add Wald Statistic for Clustered Standard Errors
In case you use clustered standard errors you should not use R2 to compute the F - Statistic. What one does instead is to compute a Wald Statistic. Note, the necessary number of degrees of freedom is the number of available clusters. Clustered standard errors do converge with an increase in the number of clusters and not with an increase in the number of observations.